+++ src/renderer/store/ops.ts
@@ -14,6 +14,7 @@ import {
 
 export const DELETE_EVERY_SECOND_WORD = 'DELETE_EVERY_SECOND_WORD';
 export const UNDO_DELETE_EVERY_SECOND_WORD = 'UNDO_DELETE_EVERY_SECOND_WORD';
+
 export const CHANGE_WORD_TO_SWAMP = 'CHANGE_WORD_TO_SWAMP';
 export const UNDO_CHANGE_WORD_TO_SWAMP = 'UNDO_CHANGE_WORD_TO_SWAMP';
