+++ src/renderer/store/reducers/currentProjectReducer.ts
@@ -1,0 +1,63 @@
+import { Project, Transcription, Word } from 'sharedTypes';
+import {
+  CURRENT_PROJECT_CLOSED,
+  PROJECT_CREATED,
+  PROJECT_OPENED,
+  TRANSCRIPTION_CREATED,
+} from '../actions';
+import { Action, ApplicationStore, initialStore } from '../helpers';
+import {
+  ChangeWordToSwampPayload,
+  UndoChangeWordToSwampPayload,
+  UndoDeleteEverySecondWordPayload,
+} from '../opPayloads';
+import {
+  CHANGE_WORD_TO_SWAMP,
+  DELETE_EVERY_SECOND_WORD,
+  UNDO_CHANGE_WORD_TO_SWAMP,
+  UNDO_DELETE_EVERY_SECOND_WORD,
+} from '../ops';
+import transcriptionReducer from './transcriptionReducer';
+
+const currentProjectReducer: (
+  currentProject: ApplicationStore['currentProject'],
+  action: Action<any>
+) => ApplicationStore['currentProject'] = (
+  currentProject = initialStore.currentProject,
+  action
+) => {
+  if (action.type === PROJECT_CREATED || action.type === PROJECT_OPENED) {
+    return action.payload as Project;
+  }
+
+  if (action.type === CURRENT_PROJECT_CLOSED) {
+    return null;
+  }
+
+  if (action.type === TRANSCRIPTION_CREATED && currentProject !== null) {
+    return {
+      ...currentProject,
+      transcription: action.payload as Transcription,
+    };
+  }
+
+  if (
+    [
+      DELETE_EVERY_SECOND_WORD,
+      UNDO_DELETE_EVERY_SECOND_WORD,
+      CHANGE_WORD_TO_SWAMP,
+      UNDO_CHANGE_WORD_TO_SWAMP,
+    ].includes(action.type) &&
+    currentProject !== null &&
+    currentProject.transcription !== null
+  ) {
+    return {
+      ...currentProject,
+      transcription: transcriptionReducer(currentProject.transcription, action),
+    };
+  }
+
+  return currentProject;
+};
+
+export default currentProjectReducer;
